+++ src/components/Door.hpp
@@ -48,7 +48,7 @@ public:
 
     BitmapDoor(YAML::Node data);
     void update(RenderManager* renderer, int dt) override;
-    bool clicked(int x, int y) override;
+    bool clicked(float x, float y) override;
 
 };
 
@@ -63,6 +63,6 @@ public:
 
     SpriteDoor(YAML::Node data, AssetManager* assets);
     void update(RenderManager* renderer, int dt) override;
-    bool clicked(int x, int y) override;
+    bool clicked(float x, float y) override;
 
 };
