+++ src/managers/src/RoomManager.cpp
@@ -81,10 +81,10 @@ void RoomManager::update(RenderManager* renderer, int dt)
 }
 
 
-bool RoomManager::walkable(int x, int y) { return active_room_->walkable(x, y); }
-GameObject* RoomManager::get_object(int x, int y) { return active_room_->get_object(x, y); }
-GameObject* RoomManager::get_item(int x, int y) { return active_room_->get_item(x, y); }
-GameObject* RoomManager::get_hot_spot(int x, int y) { return active_room_->get_hot_spot(x, y); }
+bool RoomManager::walkable(float x, float y) { return active_room_->walkable(x, y); }
+GameObject* RoomManager::get_object(float x, float y) { return active_room_->get_object(x, y); }
+GameObject* RoomManager::get_item(float x, float y) { return active_room_->get_item(x, y); }
+GameObject* RoomManager::get_hot_spot(float x, float y) { return active_room_->get_hot_spot(x, y); }
 void RoomManager::remove_item(std::string id) { active_room_->remove_item(id); }
