+++ src/math/Visual.hpp
@@ -3,6 +3,6 @@
 struct Visual
 {
     float scale {1.0f};
-    int dx {0};
-    int dy {0};
+    float dx {0};
+    float dy {0};
 };
